Adding Blinn-Phong Lighting Example.

+++ engine/include/events/enums/event_category.h
@@ -3,7 +3,7 @@
 #include "vlkypch.h"
 
 namespace Vulkyrie::Events {
-    enum class EventCategory : u8 {
+    enum class EventCategory : u32 {
         ApplicationEvent = 1 << 0, // Application Event.
         Input = 1 << 1,            // Input event.
         Keyboard = 1 << 2,         // Keyboard event.
