+++ engine/include/physics/shape_type.h
@@ -3,7 +3,7 @@
 #include "vlkypch.h"
 
 namespace Vulkyrie::Physics {
-    enum class ShapeType : u8 {
+    enum class ShapeType : u32 {
         Box = 0,
         Sphere = 1,
         Capsule = 2,
